Revert "include: Add a copy of the NFC kernel header"
This reverts commit fb7664e2995f2b5bed4b93c7c82aab14c2485eb8 and changes
several other files to match the scope of revert (using back system
header, instead of local copy).

The reverted commit argued it is better to miss header updates than not
being able to run the neard daemon at all.  It's not correct.  The
/usr/include/linux/nfc.h header does not affect the runnability of
daemon but the build stage.  The daemon, once compiled, will be able to
run even if nfc.h system header changes.

However in case of any important differences in nfc.h, it is better to
fail early during build time, than later during runtime because of
differences of ABI (structures or functions) between kernel and
user-space.

Rely back on system/kernel header, to avoid manual synchronisation.
This will also force us to keep the neard updated if kernel header
changes.

Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>